+++ dal-jpa/src/main/java/cn/tannn/jdevelops/demo/jpa/service/impl/UserServiceImpl.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Service;
  */
 @Slf4j
 @Service
-public class UserServiceImpl extends J2ServiceImpl<UserDao, User, Integer> implements UserService {
+public class UserServiceImpl extends J2ServiceImpl<UserDao, User, Long> implements UserService {
 
     public UserServiceImpl() {
         super(User.class);
